move set_up_ious
